v3.5. Stage slider nice. PMT and mouse_click+s implemented in scan.

+++ nionswift_plugin/stage/stage_inst.py
@@ -102,7 +102,7 @@ class stageDevice(Observable.Observable):
 
     @slider_range_f.setter
     def slider_range_f(self, value):
-        self.__slider_range=value*100
+        self.__slider_range=int(value*100)
         if self.__slider_range<32000:
             try:
                 self.slider_event.fire()
